Really unconvinced these constants belong outside of their sole usage
