+++ src/wikisearch.ts
@@ -3,7 +3,110 @@ import { MessageEmbed } from "discord.js";
 import { Effect, Familiar, Item, Monster, Skill, Thing } from "./things";
 import { KOLClient } from "./kol";
 import { cleanString } from "./utils";
-import { PACKAGES, REVERSE_PACKAGES } from "./constants";
+
+const PACKAGES = new Map([
+  ["iceberglet", "ice pick"],
+  ["great ball of frozen fire", "evil flaming eyeball pendant"],
+  ["naughty origami kit", "naughty paper shuriken"],
+  ["packet of mayfly bait", "mayfly bait necklace"],
+  ["container of spooky putty", "spooky putty sheet"],
+  ["stinky cheese ball", "stinky cheese diaper"],
+  ["grumpy bumpkin's pumpkin seed catalog", "packet of pumpkin seeds"],
+  ["make-your-own-vampire-fangs kit", "plastic vampire fangs"],
+  ["mint salton pepper's peppermint seed catalog", "peppermint pip packet"],
+  ["pete & jackie's dragon tooth emporium catalog", "packet of dragon's teeth"],
+  ["folder holder", "over-the-shoulder folder holder"],
+  ["discontent™ winter garden catalog", "packet of winter seeds"],
+  ["ed the undying exhibit crate", "the crown of ed the undying"],
+  ["pack of every card", "deck of every card"],
+  ["diy protonic accelerator kit", "protonic accelerator pack"],
+  ["dear past self package", "time-spinner"],
+  ["suspicious package", "kremlin's greatest briefcase"],
+  ["li-11 motor pool voucher", "asdon martin keyfob"],
+  ["corked genie bottle", "genie bottle"],
+  ["pantogram", "portable pantogram"],
+  ["locked mumming trunk", "mumming trunk"],
+  ["january's garbage tote (unopened)", "january's garbage tote"],
+  ["pokéfam guide to capturing all of them", "packet of tall grass seeds"],
+  ["songboom™ boombox box", "songboom™ boombox"],
+  ["bastille batallion control rig crate", "bastille batallion control rig"],
+  ["latte lovers club card", "latte lovers member's mug"],
+  ["kramco industries packing carton", "kramco sausage-o-matic™"],
+  ["mint condition lil' doctor™ bag", "lil' doctor™ bag"],
+  ["vampyric cloake pattern", "vampyric cloake"],
+  ["fourth of may cosplay saber kit", "fourth of may cosplay saber"],
+  ["rune-strewn spoon coccoon", "hewn moon-rune spoon"],
+  ["beach comb box", "beach comb"],
+  ["unopened eight days a week pill keeper", "eight days a week pill keeper"],
+  ["unopened diabolic pizza cube box", "diabolic pizza cube"],
+  ["mint-in-box powerful glove", "powerful glove"],
+  ["better shrooms and gardens catalog", "packet of mushroom spores"],
+  ["guzzlr application", "guzzlr tablet"],
+  ["bag of iunion stones", "iunion crown"],
+  ["packaged spinmaster™ lathe", "spinmaster™ lathe"],
+  ["bagged cargo cultist shorts", "cargo cultist shorts"],
+  ["packaged knock-off retro superhero cape", "unwrapped knock-off retro superhero cape"],
+  ["box o' ghosts", "greedy ghostling"],
+  ["packaged miniature crystal ball", "miniature crystal ball"],
+  ["emotion chip", "spinal-fluid-covered emotion chip"],
+  ["power seed", "potted power plant"],
+  ["packaged backup camera", "backup camera"],
+  ["packaged familiar scrapbook", "familiar scrapbook"],
+  ["packaged industrial fire extinguisher", "industrial fire extinguisher"],
+  ["packaged daylight shavings helmet", "daylight shavings helmet"],
+  ["packaged cold medicine cabinet", "cold medicine cabinet"],
+  ["undrilled cosmic bowling ball", "cosmic bowling ball"],
+  ["combat lover's locket lockbox", "combat lover's locket"],
+  ["undamaged unbreakable umbrella", "unbreakable umbrella"],
+  ["retrospecs try-at-home kit", "retrospecs"],
+  ["fresh can of paint", "fresh coat of paint"],
+  ["mint condition magnifying glass", "cursed magnifying glass"],
+  ["packaged june cleaver", "june cleaver"],
+  ["designer sweatpants (new old stock)", "designer sweatpants"],
+  ["unopened tiny stillsuit", "tiny stillsuit"],
+  ["packaged jurassic parka", "jurassic parka"],
+  ["boxed autumn-aton", "autumn-aton"],
+]);
+
+const ghostlings: [string, string][] = [
+  ["grinning ghostling", "box o' ghosts"],
+  ["gregarious ghostling", "box o' ghosts"],
+  ["greedy ghostling", "box o' ghosts"],
+];
+
+const foldables: [string, string][] = [
+  ["ice baby", "iceberglet"],
+  ["ice pick", "iceberglet"],
+  ["ice skates", "iceberglet"],
+  ["ice sickle", "iceberglet"],
+  ["liar's pants", "great ball of frozen fire"],
+  ["flaming juggler's balls", "great ball of frozen fire"],
+  ["flaming pink shirt", "great ball of frozen fire"],
+  ["flaming familiar doppelgänger", "great ball of frozen fire"],
+  ["evil flaming eyeball pendant", "great ball of frozen fire"],
+  ["naughty paper shuriken", "naughty origami kit"],
+  ["origami pasties", "naughty origami kit"],
+  ["origami riding crop", "naughty origami kit"],
+  ['origami "gentlemen\'s" magazine', "naughty origami kit"],
+  ["naughty fortune teller", "naughty origami kit"],
+  ["spooky putty mitre", "container of spooky putty"],
+  ["spooky putty leotard", "container of spooky putty"],
+  ["spooky putty ball", "container of spooky putty"],
+  ["spooky putty sheet", "container of spooky putty"],
+  ["spooky putty snake", "container of spooky putty"],
+  ["stinky cheese sword", "stinky cheese ball"],
+  ["stinky cheese diaper", "stinky cheese ball"],
+  ["stinky cheese wheel", "stinky cheese ball"],
+  ["stinky cheese eye", "stinky cheese ball"],
+  ["staff of queso escusado", "stinky cheese ball"],
+];
+
+const reversed: [string, string][] = Array.from(PACKAGES.keys()).map((key) => [
+  PACKAGES.get(key) || "",
+  key,
+]);
+
+const REVERSE_PACKAGES = new Map(reversed.concat(ghostlings).concat(foldables));
 
 type FoundName = {
   name: string;
